fix: Cascade delete optimized

+++ pkg/usecase/usecase/resource/network.go
@@ -135,11 +135,7 @@ func (nuc *networkUseCase) Delete(ctx context.Context, subnetwork *model.Network
 	}
 	req := ctx.Value(context.RequestKey).(dto.DeleteNetworkRequest)
 
-	project, errRepo := nuc.projectRepo.Find(ctx, option.Option{Value: repository.FindProjectByIDRequest{ID: ctx.Value(context.ProjectIDKey).(string)}})
-    if !errRepo.IsOk() {
-        return errRepo
-    }
-    ctx.Set(context.CurrentNamespace, project.Namespace)
+	
 
 	foundNetwork, errNet := repo.FindNetwork(ctx, option.Option{Value: resourceRepo.FindResourceOption{Name: req.IdentifierID}})
 	if !errNet.IsOk() {
@@ -147,6 +143,11 @@ func (nuc *networkUseCase) Delete(ctx context.Context, subnetwork *model.Network
 	}
 	*subnetwork = *foundNetwork
 	if req.Cascade {
+		project, errRepo := nuc.projectRepo.Find(ctx, option.Option{Value: repository.FindProjectByIDRequest{ID: ctx.Value(context.ProjectIDKey).(string)}})
+		if !errRepo.IsOk() {
+			return errRepo
+		}
+		ctx.Set(context.CurrentNamespace, project.Namespace)
 		if errCascade := repo.DeleteNetworkCascade(ctx, foundNetwork); !errCascade.IsOk() {
 			return errCascade
 		}
